feat: layout component type docs
