+++ .changeset/fast-buses-help.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/media-object': patch
+'@twilio-paste/core': patch
+---
+
+[Media Object]: Improved type annotations
