+++ .changeset/grumpy-boxes-pay.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/flex': patch
+'@twilio-paste/core': patch
+---
+
+[Flex]: Improved TsDoc annotations
