+++ .changeset/healthy-books-kneel.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/aspect-ratio': patch
+'@twilio-paste/core': patch
+---
+
+[Aspect Ratio]: improved TsDoc annotations
