+++ .changeset/neat-papayas-play.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/grid': patch
+'@twilio-paste/core': patch
+---
+
+[Grid]: Improved TsDoc annotations
