+++ .changeset/orange-flowers-sip.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/codemods': patch
+'@twilio-paste/core': patch
+---
+
+[Codemods] updated exports map
