+++ .changeset/poor-ways-collect.md
@@ -1,0 +1,6 @@
+---
+'@twilio-paste/stack': patch
+'@twilio-paste/core': patch
+---
+
+[Stack]: Improved TsDoc annotations
