+++ packages/paste-codemods/tools/.cache/mappings.json
@@ -307,9 +307,6 @@
   "MediaFigure": "@twilio-paste/core/media-object",
   "MediaObject": "@twilio-paste/core/media-object",
   "Stack": "@twilio-paste/core/stack",
-  "getStackChildMargins": "@twilio-paste/core/stack",
-  "getStackDisplay": "@twilio-paste/core/stack",
-  "getStackStyles": "@twilio-paste/core/stack",
   "BOX_PROPS_TO_BLOCK": "@twilio-paste/core/box",
   "Box": "@twilio-paste/core/box",
   "StyledBox": "@twilio-paste/core/box",
