+++ packages/paste-codemods/tools/__tests__/tools.spec.ts
@@ -42,9 +42,6 @@ describe('generatePackageExportsMap', () => {
       Stack: '@twilio-paste/core/stack',
       StyledBox: '@twilio-paste/core/box',
       getCustomElementStyles: '@twilio-paste/core/box',
-      getStackChildMargins: '@twilio-paste/core/stack',
-      getStackDisplay: '@twilio-paste/core/stack',
-      getStackStyles: '@twilio-paste/core/stack',
       safelySpreadBoxProps: '@twilio-paste/core/box',
       useComboboxPrimitive: '@twilio-paste/core/combobox-primitive',
       useDisclosurePrimitiveState: '@twilio-paste/core/disclosure-primitive',
