+++ packages/paste-core/layout/aspect-ratio/type-docs.json
@@ -1,0 +1,11 @@
+{
+  "AspectRatio": {
+    "ratio": {
+      "type": "string",
+      "defaultValue": null,
+      "required": true,
+      "externalProp": false,
+      "description": "Determines the aspect ratio of the element. Use a colon separated number pattern (width:height)."
+    }
+  }
+}
