+++ packages/paste-core/layout/flex/package.json
@@ -18,8 +18,9 @@
     "dist"
   ],
   "scripts": {
-    "build": "yarn clean && NODE_ENV=production node build.js && tsc",
+    "build": "yarn clean && NODE_ENV=production node build.js && tsc && yarn build:typedocs",
     "build:js": "NODE_ENV=development node build.js",
+    "build:typedocs": "tsx ../../../../tools/build/generate-type-docs",
     "clean": "rm -rf ./dist",
     "tsc": "tsc"
   },
@@ -52,6 +53,7 @@
     "@types/react-dom": "^18.0.10",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "tsx": "^3.12.10",
     "typescript": "^4.9.4"
   }
 }
