+++ packages/paste-core/layout/media-object/src/index.tsx
@@ -1,3 +1,6 @@
-export * from './MediaObject';
-export * from './MediaFigure';
-export * from './MediaBody';
+export {MediaObject} from './MediaObject';
+export type {MediaObjectProps} from './MediaObject';
+export {MediaFigure} from './MediaFigure';
+export type {MediaFigureProps} from './MediaFigure';
+export {MediaBody} from './MediaBody';
+export type {MediaBodyProps} from './MediaBody';
