+++ packages/paste-website/src/pages/components/flex/api.mdx
@@ -1,0 +1,56 @@
+export const meta = {
+  title: 'Flex - API',
+  package: '@twilio-paste/flex',
+  description:
+    'Flex is a layout component that allows you to build flexbox based elements and layouts. These flexible elements and layouts allow your components to adapt to the available space or screen size.',
+  slug: '/components/flex/api',
+};
+
+import Changelog from '@twilio-paste/flex/CHANGELOG.md'; // I don't know why this is needed but if you remove it the page fails to render
+import packageJson from '@twilio-paste/flex/package.json';
+
+import {SidebarCategoryRoutes} from '../../../constants';
+import ComponentPageLayout from '../../../layouts/ComponentPageLayout';
+import {getFeature, getNavigationData, getComponentApi} from '../../../utils/api';
+
+export default ComponentPageLayout;
+
+export const getStaticProps = async () => {
+  const navigationData = await getNavigationData();
+  const feature = await getFeature('Flex');
+  const {componentApi, componentApiTocData} = getComponentApi('@twilio-paste/flex');
+  return {
+    props: {
+      data: {
+        ...packageJson,
+        ...feature,
+      },
+      componentApi,
+      mdxHeadings: [...mdxHeadings, ...componentApiTocData],
+      navigationData,
+      pageHeaderData: {
+        categoryRoute: SidebarCategoryRoutes.COMPONENTS,
+        githubUrl: 'https://github.com/twilio-labs/paste/tree/main/packages/paste-core/layout/flex',
+        storybookUrl: '/?path=/story/layout-flex--flex-options',
+      },
+    },
+  };
+};
+
+## Installation
+
+```bash
+yarn add @twilio-paste/flex - or - yarn add @twilio-paste/core
+```
+
+## Usage
+
+```jsx
+import {Flex} from '@twilio-paste/core/flex';
+
+const Component = () => <Flex>Foo</Flex>;
+```
+
+## Props
+
+<PropsTable componentApi={props.componentApi} />
