+++ packages/paste-website/src/pages/components/grid/api.mdx
@@ -1,0 +1,59 @@
+export const meta = {
+  title: 'Grid - API',
+  package: '@twilio-paste/grid',
+  description: 'The grid component controls the visual layout and spacing between elements on a page or in containers.',
+  slug: '/components/grid/api',
+};
+
+import Changelog from '@twilio-paste/grid/CHANGELOG.md'; // I don't know why this is needed but if you remove it the page fails to render
+import packageJson from '@twilio-paste/grid/package.json';
+
+import {SidebarCategoryRoutes} from '../../../constants';
+import ComponentPageLayout from '../../../layouts/ComponentPageLayout';
+import {getFeature, getNavigationData, getComponentApi} from '../../../utils/api';
+
+export default ComponentPageLayout;
+
+export const getStaticProps = async () => {
+  const navigationData = await getNavigationData();
+  const feature = await getFeature('Grid');
+  const {componentApi, componentApiTocData} = getComponentApi('@twilio-paste/grid');
+  return {
+    props: {
+      data: {
+        ...packageJson,
+        ...feature,
+      },
+      componentApi,
+      mdxHeadings: [...mdxHeadings, ...componentApiTocData],
+      navigationData,
+      pageHeaderData: {
+        categoryRoute: SidebarCategoryRoutes.COMPONENTS,
+        githubUrl: 'https://github.com/twilio-labs/paste/tree/main/packages/paste-core/layout/grid',
+        storybookUrl: '/?path=/story/layout-grid--grid-12-column-and-gutter-support',
+      },
+    },
+  };
+};
+
+## Installation
+
+```bash
+yarn add @twilio-paste/grid - or - yarn add @twilio-paste/core
+```
+
+## Usage
+
+```jsx
+import {Grid, Column} from '@twilio-paste/core/grid';
+
+const Component = () => (
+  <Grid>
+    <Column>Foo</Column>
+  </Grid>
+);
+```
+
+## Props
+
+<PropsTable componentApi={props.componentApi} />
