+++ packages/paste-website/src/pages/components/media-object/changelog.mdx
@@ -1,0 +1,38 @@
+export const meta = {
+  title: 'Media Object - Layout',
+  package: '@twilio-paste/media-object',
+  description:
+    'Media Object is a layout utility that allows an image or figure to be positioned horizontally next to some form of content.',
+  slug: '/components/media-object/changelog',
+};
+
+import Changelog from '@twilio-paste/media-object/CHANGELOG.md';
+import packageJson from '@twilio-paste/media-object/package.json';
+
+import {SidebarCategoryRoutes} from '../../../constants';
+import ComponentPageLayout from '../../../layouts/ComponentPageLayout';
+import {getFeature, getNavigationData} from '../../../utils/api';
+
+export default ComponentPageLayout;
+
+export const getStaticProps = async () => {
+  const navigationData = await getNavigationData();
+  const feature = await getFeature('Media Object');
+  return {
+    props: {
+      data: {
+        ...packageJson,
+        ...feature,
+      },
+      navigationData,
+      mdxHeadings,
+      pageHeaderData: {
+        categoryRoute: SidebarCategoryRoutes.COMPONENTS,
+        githubUrl: 'https://github.com/twilio-labs/paste/tree/main/packages/paste-core/layout/media-object',
+        storybookUrl: '/?path=/story/layout-media-object--default',
+      },
+    },
+  };
+};
+
+<Changelog />
