+++ packages/paste-website/src/pages/components/stack/api.mdx
@@ -1,0 +1,57 @@
+export const meta = {
+  title: 'Stack - Layout',
+  package: '@twilio-paste/stack',
+  description: 'Stack is a horizontal and vertical flow component used to create space between components.',
+  slug: '/components/stack/api',
+};
+
+import Changelog from '@twilio-paste/stack/CHANGELOG.md'; // I don't know why this is needed but if you remove it the page fails to render
+import packageJson from '@twilio-paste/stack/package.json';
+
+import {SidebarCategoryRoutes} from '../../../constants';
+import ComponentPageLayout from '../../../layouts/ComponentPageLayout';
+import {getFeature, getNavigationData, getComponentApi} from '../../../utils/api';
+
+export default ComponentPageLayout;
+
+export const getStaticProps = async () => {
+  const navigationData = await getNavigationData();
+  const feature = await getFeature('Stack');
+  const {componentApi, componentApiTocData} = getComponentApi('@twilio-paste/stack');
+  return {
+    props: {
+      data: {
+        ...packageJson,
+        ...feature,
+      },
+      componentApi,
+      mdxHeadings: [...mdxHeadings, ...componentApiTocData],
+      navigationData,
+      pageHeaderData: {
+        categoryRoute: SidebarCategoryRoutes.COMPONENTS,
+        githubUrl: 'https://github.com/twilio-labs/paste/tree/main/packages/paste-core/layout/stack',
+        storybookUrl: '/?path=/story/layout-stack--default-stack',
+      },
+    },
+  };
+};
+
+## Installation
+
+```bash
+yarn add @twilio-paste/stack
+```
+
+## Usage
+
+```jsx
+import {Stack} from '@twilio-paste/core/stack';
+
+<Stack orientation="vertical" spacing="space40">
+  Foo
+</Stack>;
+```
+
+## Props
+
+<PropsTable componentApi={props.componentApi} />
