+++ packages/paste-website/src/pages/components/stack/changelog.mdx
@@ -1,0 +1,37 @@
+export const meta = {
+  title: 'Stack - Layout',
+  package: '@twilio-paste/stack',
+  description: 'Stack is a horizontal and vertical flow component used to create space between components.',
+  slug: '/components/stack/changelog',
+};
+
+import Changelog from '@twilio-paste/stack/CHANGELOG.md';
+import packageJson from '@twilio-paste/stack/package.json';
+
+import {SidebarCategoryRoutes} from '../../../constants';
+import ComponentPageLayout from '../../../layouts/ComponentPageLayout';
+import {getFeature, getNavigationData} from '../../../utils/api';
+
+export default ComponentPageLayout;
+
+export const getStaticProps = async () => {
+  const navigationData = await getNavigationData();
+  const feature = await getFeature('Stack');
+  return {
+    props: {
+      data: {
+        ...packageJson,
+        ...feature,
+      },
+      navigationData,
+      mdxHeadings,
+      pageHeaderData: {
+        categoryRoute: SidebarCategoryRoutes.COMPONENTS,
+        githubUrl: 'https://github.com/twilio-labs/paste/tree/main/packages/paste-core/layout/stack',
+        storybookUrl: '/?path=/story/layout-stack--default-stack',
+      },
+    },
+  };
+};
+
+<Changelog />
